requirements.txt
flask==2.3.3 
ultralytics==8.0.186 
opencv-python==4.8.1.78 
pillow==10.0.0 
numpy==1.24.3 
easyocr
torch
